chore: rename lib/gui/string_utils.h => lib/gui/StringUtils.h
