+++ src/apps/deskflow-gui/MainWindow.cpp
@@ -20,9 +20,9 @@
 #include "gui/Diagnostic.h"
 #include "gui/Logger.h"
 #include "gui/Messages.h"
+#include "gui/StringUtils.h"
 #include "gui/core/CoreProcess.h"
 #include "gui/ipc/DaemonIpcClient.h"
-#include "gui/string_utils.h"
 #include "gui/style_utils.h"
 #include "gui/styles.h"
 #include "net/FingerprintDatabase.h"
