+++ src/lib/gui/StringUtils.h
